* sendto_kdc.c (krb5_sendto_kdc): Get rid of unecessary casts
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@14041 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/os/sendto_kdc.c
@@ -178,7 +178,7 @@ krb5_sendto_kdc (context, message, realm, reply, use_master)
 	    fprintf (stderr, "sending...");
 #endif
 	    if (send(socklist[host],
-		       message->data, (int) message->length, 0) 
+		       message->data, message->length, 0) 
 		!= message->length) {
 #ifdef DEBUG
 		perror ("sendto");
@@ -205,7 +205,7 @@ krb5_sendto_kdc (context, message, realm, reply, use_master)
 		    goto out;
 		}
 		if ((cc = recv(socklist[host],
-			       reply->data, (int) reply->length, 
+			       reply->data, reply->length, 
 			       0)) == SOCKET_ERROR)
 		  {
 		    /* man page says error could be:
